created controller tests that deleting saved places returns correct status #300

+++ src/main/java/ch/uzh/ifi/hase/soprafs26/controller/SavedPlaceController.java
@@ -86,10 +86,16 @@ public class SavedPlaceController {
     @DeleteMapping("/users/{userId}/savedplaces/{savedPlaceId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteSavedPlace(
+        @PathVariable Long userId,
         @PathVariable Long savedPlaceId,
         @RequestHeader (value = "Authorization", required = false) String token
     ){
-        userService.validateToken(token);
+        User loggedInUser = userService.validateToken(token);
+        if (!loggedInUser.getId().equals(userId)) {
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, 
+				"You are not allowed to get other users' saved places.");
+		}
+
         savedPlaceService.deleteSavedPlace(savedPlaceId, token);
     }
